Fix Non-Retina App Icons
There was a graphics glitch on Monterey on non-Retina displays.
The edges appeared light blue on white backgrounds. Strangely, this only affected the app icon, even though the icns file itself looked fine.
It could be something related to pre-multiplied alpha channel, but I'm not sure.
The only way I got the icon to render properly was to do the following:

- I re-did the icons in Sketch
- I removed a white background from the mask
- I replaced the icns file with an icon in the Asset Catalog