Add preference for hiding the menu bar icon
Some users prefer a clean menu bar with fewer icons.

I added a checkbox that's bound to the isVisible property of the status item.
The checkbox should automatically be disabled on macOS < 10.12 (don't support isVisible).

However, just hiding the status item would lead to an app without any UI that the user can't quit!
So I added logic that shows the preference window when the user double-clicks the app in the finder.
I also added a "Open Web UI" and a "Quit" button so that the user can interact with the app when the menu item is hidden.

I also updated the preference view to use auto layout, because it was sometimes clipped.