Add appcast.xml generation action
This adds an action to generate the appcast.xml and push it to cloud
storage bucket. This replaces the thing currently running on
build.syncthing.net, which is in a slow death spiral.

I set the action to run nightly and manually; since it's based on the
published releases it should be fine to run periodically (generating
identical contents each time) and manually in case a quick update is
required.

(Also refreshed the dependencies for the generation script while I was
in there.)